% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{install_all}
\alias{install_all}
\title{Install all packages on clipboard}
\usage{
install_all(pkgs = NULL)
}
\description{
install all packages in the package list rendered by \link[devtools]{install_github} on the clipboard
}
