% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.function.R
\name{jobs.function}
\alias{jobs.function}
\title{Title}
\usage{
jobs.function(eqtls, eqtls_se)
}
\arguments{
\item{eqtls}{}

\item{eqtls_se}{}
}
\description{
Title
}
\examples{
jobs.function(eqtls,eqtls_se)
}
