% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.functions.R
\name{jobs.nnls.weights}
\alias{jobs.nnls.weights}
\title{Title}
\usage{
jobs.nnls.weights(beta, se)
}
\arguments{
\item{beta}{This matrix contains eQTL effect sizes for both bulk and single-cell data. Column 1: Gene-SNP pair identifiers. Column 2: Bulk effect size. Columns 3+: Cell type-specific eQTL effect sizes.}

\item{se}{This matrix has the same dimensions as the effect size matrix and represents the standard errors associated with the eQTL effect sizes. Column 1: Gene-SNP pair identifiers (should match the effect size matrix). Column 2: Bulk effect size standard error. Columns 3+: Standard errors for the cell type-specific eQTL effect sizes.}
}
\description{
Title
}
\examples{
jobs.nnls.weights(beta,se,weight)
}
