% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{jobscript}
\alias{jobscript}
\title{Write a code chunk to file}
\usage{
jobscript(code, filename, dir = NULL, workingDir = getwd(), ...)
}
\arguments{
\item{code}{\code{expression} to be written to file. Either one created with \link[rlang]{expr}, or provided wrapped in \code{\{\}}}

\item{filename}{\code{character} filename}

\item{dir}{\code{character} vector of directory path(s)}

\item{workingDir}{The working directory in which to run the job. When
\code{NULL} (the default), the parent directory of the R script is used.}

\item{...}{
  Arguments passed on to \code{\link[rstudioapi:jobRunScript]{rstudioapi::jobRunScript}}
  \describe{
    \item{\code{path}}{The path to the R script to be run.}
    \item{\code{name}}{A name for the background job. When \code{NULL} (the default),
the filename of the script is used as the job name.}
    \item{\code{encoding}}{The text encoding of the script, if known.}
    \item{\code{importEnv}}{Whether to import the global environment into the job.}
    \item{\code{exportEnv}}{The name of the environment in which to export the R
objects created by the job. Use \code{""} (the default) to skip export,
\code{"R_GlobalEnv"}` to export to the global environment, or the name of an
environment object to create an object with that name.}
  }}
}
\value{
\code{character} the path to the file
}
\description{
Write the \code{code} provided to a file indicated by \code{filename} in the directory indicated by \code{dir}
}
